Grab correct token
